Implement reconstruct_event as a separated function.
Re-organize reco directory.
Remove unused old mc functions.
Cosmetics in prints and logger.
Update tests to new nexus persistency and FANAL re-arrangemaent
