+++ fanal/reco/reco_functions_test.py
@@ -1,8 +1,8 @@
 """
-Tests for position
+Tests for reco functions
 """
 
-import numpy  as np
+import numpy as np
 import pandas as pd
 import pytest
 
@@ -22,21 +22,28 @@ from invisible_cities.core.testing_utils import float_arrays
 from invisible_cities.core.testing_utils import FLOAT_ARRAY
 from invisible_cities.core.testing_utils import random_length_float_arrays
 
-import invisible_cities.core.system_of_units as units
-from   invisible_cities.evm.event_model      import MCHit, Voxel
+import invisible_cities.core.system_of_units       as units
+from   invisible_cities.evm.event_model        import MCHit
+from   invisible_cities.evm.event_model        import Voxel
 
-from fanal.core.fanal_types import DetName
-from fanal.reco.position    import translate_hit_positions
-from fanal.reco.position    import check_event_fiduciality
+from fanal.core.fanal_types    import DetName
+
+from fanal.reco.reco_functions import get_mc_energy
+from fanal.reco.reco_functions import smear_evt_energy
+from fanal.reco.reco_functions import S1_Eth, S1_WIDTH, EVT_WIDTH
+
+from fanal.reco.reco_functions import translate_hit_positions
+from fanal.reco.reco_functions import check_event_fiduciality
 
 
+############################  POSITION FUNCTIONS  ############################
 
 def test_translate_hit_positions():
     # Check of assymetric detector
-    mcHits = [{'label': 'test', 'time':  0 * units.mus, 'x': 0., 'y': 0., 'z': 0., 'E': 0.01},
-              {'label': 'test', 'time': 10 * units.mus, 'x': 0., 'y': 0., 'z': 0., 'E': 0.01},
-              {'label': 'test', 'time': 20 * units.mus, 'x': 0., 'y': 0., 'z': 0., 'E': 0.01}]
-    mcHits = pd.DataFrame(mcHits, columns = ('label', 'time', 'x', 'y', 'z', 'E'))
+    mcHits = [{'label': 'test', 'time':  0 * units.mus, 'x': 0., 'y': 0., 'z': 0., 'energy': 0.01},
+              {'label': 'test', 'time': 10 * units.mus, 'x': 0., 'y': 0., 'z': 0., 'energy': 0.01},
+              {'label': 'test', 'time': 20 * units.mus, 'x': 0., 'y': 0., 'z': 0., 'energy': 0.01}]
+    mcHits = pd.DataFrame(mcHits, columns = ('label', 'time', 'x', 'y', 'z', 'energy'))
 
     drift_velocity = 1. * units.mm / units.mus
     translate_hit_positions(DetName.next100, mcHits, drift_velocity)
@@ -46,10 +53,10 @@ def test_translate_hit_positions():
     assert_array_equal(OK_shifted_z, mcHits.shifted_z)
 
     # Check of ssymetric detectors
-    mcHits = [{'label': 'test', 'time':  0 * units.mus, 'x': 0., 'y': 0., 'z':    0., 'E': 0.01},
-              {'label': 'test', 'time': 40 * units.mus, 'x': 0., 'y': 0., 'z':  500., 'E': 0.01},
-              {'label': 'test', 'time': 20 * units.mus, 'x': 0., 'y': 0., 'z': -500., 'E': 0.01}]
-    mcHits = pd.DataFrame(mcHits, columns = ('label', 'time', 'x', 'y', 'z', 'E'))
+    mcHits = [{'label': 'test', 'time':  0 * units.mus, 'x': 0., 'y': 0., 'z':    0., 'energy': 0.01},
+              {'label': 'test', 'time': 40 * units.mus, 'x': 0., 'y': 0., 'z':  500., 'energy': 0.01},
+              {'label': 'test', 'time': 20 * units.mus, 'x': 0., 'y': 0., 'z': -500., 'energy': 0.01}]
+    mcHits = pd.DataFrame(mcHits, columns = ('label', 'time', 'x', 'y', 'z', 'energy'))
 
     drift_velocity = 1. * units.mm / units.mus
     translate_hit_positions(DetName.next500, mcHits, drift_velocity)
@@ -124,4 +131,49 @@ def test_check_event_fiduciality():
 
 
 
+############################   ENERGY FUNCTIONS   ############################
 
+def test_get_mc_energy():
+    # Collection of hits within 1 S1
+    mcHits = [{'label': 'test', 'time': 1 * units.ns, 'x': 0., 'y': 0., 'z': 0., 'energy': 0.10},
+              {'label': 'test', 'time': 1 * units.ns, 'x': 0., 'y': 0., 'z': 0., 'energy': 0.01},
+              {'label': 'test', 'time': 1 * units.ns, 'x': 0., 'y': 0., 'z': 0., 'energy': 0.05}]
+    mcHits = pd.DataFrame(mcHits, columns = ('label', 'time', 'x', 'y', 'z', 'energy'))
+    evt_mcE = get_mc_energy(mcHits)
+    assert evt_mcE == 0.16
+
+    # Collection of hits within 2 S1s in 1 single event recorded time
+    mcHits = [{'label': 'test', 'time': 1 * units.ns, 'x': 0., 'y': 0., 'z': 0., 'energy': 0.10},
+              {'label': 'test', 'time': 2 * units.ns, 'x': 0., 'y': 0., 'z': 0., 'energy': 0.01},
+              {'label': 'test', 'time': (S1_WIDTH + 1) * units.ns, 'x': 0., 'y': 0., 'z': 0., 'energy': 0.05}]
+    mcHits = pd.DataFrame(mcHits, columns = ('label', 'time', 'x', 'y', 'z', 'energy'))
+    evt_mcE = get_mc_energy(mcHits)
+    assert evt_mcE == 0.0
+
+    # Collection of hits within 2 S1s in >1 single event recorded time
+    mcHits = [{'label': 'test', 'time': 1 * units.ns, 'x': 0., 'y': 0., 'z': 0., 'energy': 0.10},
+              {'label': 'test', 'time': 2 * units.ns, 'x': 0., 'y': 0., 'z': 0., 'energy': 0.01},
+              {'label': 'test', 'time': (EVT_WIDTH + 1) * units.ns, 'x': 0., 'y': 0., 'z': 0., 'energy': 0.05}]
+    mcHits = pd.DataFrame(mcHits, columns = ('label', 'time', 'x', 'y', 'z', 'energy'))
+    evt_mcE = get_mc_energy(mcHits)
+    assert evt_mcE == 0.11
+
+    # Collection of hits within 2 S1s in 1 single event recorded time
+    # But second S1 with energy lower than threshold
+    mcHits = [{'label': 'test', 'time': 1 * units.ns, 'x': 0., 'y': 0., 'z': 0., 'energy': 0.10},
+              {'label': 'test', 'time': 2 * units.ns, 'x': 0., 'y': 0., 'z': 0., 'energy': 0.01},
+              {'label': 'test', 'time': (S1_WIDTH + 1) * units.ns,
+               'x': 0., 'y': 0., 'z': 0., 'E': S1_Eth - 5. * units.keV}]
+    mcHits = pd.DataFrame(mcHits, columns = ('label', 'time', 'x', 'y', 'z', 'energy'))
+    evt_mcE = get_mc_energy(mcHits)
+    assert evt_mcE == 0.11
+
+
+
+@flaky(max_runs=10, min_passes=9)
+def test_smear_evt_energy():
+    Qbb   = 2457.83 * units.keV
+    mc_e  = Qbb
+    sigma = 10. * units.keV
+    # 90% of times the smeared value must be in +- 2 sigmas ...
+    assert abs(smear_evt_energy(mc_e, sigma, Qbb) - mc_e) <= 2. * sigma
